REF: extract pretty_format_xml into separate module
